Works with MongoDB. Yey!

+++ ocr_machine/processor.py
@@ -1,5 +1,6 @@
 # -*- coding:utf-8 -*-
 
+from os import getenv
 import re
 from json import loads
 from ocr_machine.ocr import ocr_pipeline
@@ -9,14 +10,20 @@ from pprint import pprint
 from datetime import datetime
 from collector.settings import MONGO_URL
 from utils.dict import flatten_dict
-from pymongo import MongoClient, ASCENDING, DESCENDING, GEOSPHERE
+from pymongo import MongoClient, ASCENDING, DESCENDING, GEOSPHERE, HASHED
 
-mongo = MongoClient(MONGO_URL)
-db = mongo['bm']
-db['stations'].drop()
+db = MongoClient(MONGO_URL)['bm']
 
-create_index_one = db['stations'].ensure_index([('key', ASCENDING)], unique=True)
-create_index_two = db['stations'].ensure_index([('loc', GEOSPHERE)])
+if getenv('DROP_STATIONS') == '1': db['stations'].drop()
+
+create_index_key = db['stations'].ensure_index([('key', ASCENDING)], unique=True, cache_for=4000)
+create_index_loc = db['stations'].ensure_index([('loc', GEOSPHERE)], cache_for=4000)
+db['stations'].ensure_index([('company', ASCENDING)], cache_for=4000)
+db['stations'].ensure_index([('prices', ASCENDING)], cache_for=4000)
+db['stations'].ensure_index([('prices_last_hour', ASCENDING)], cache_for=4000)
+db['stations'].ensure_index([('prices_last_24h', ASCENDING)], cache_for=4000)
+db['stations'].ensure_index([('prices_last_yday', ASCENDING)], cache_for=4000)
+
 
 def process_station(station_as_string):
     station = loads(station_as_string, encoding='utf8')
@@ -33,59 +40,33 @@ def save_station(station):
     name = station['name']
     scraped_at = datetime.strptime(station['scraped_at'], '%Y-%m-%d %H:%M:%S')
     timestamp = int(scraped_at.timestamp())
+    hour = scraped_at.hour
+    minute = scraped_at.minute
+    day_of_the_year = scraped_at.timetuple().tm_yday
+    prices_dict = {FUEL_CODES[k]: v for k, v in station['prices'].items()}
 
     db['stations'].update_one({'key': key}, {
         "$setOnInsert": {
             'key': key,
             'name': name,
             'address': station['address'],
+            'company': station['scraper'],
             'xcode': station['xcode'],
             'xid': station['xid'],
-            'prices': {FUEL_CODES[k]: v for k, v in station['prices'].items()},
             'scraped_at': scraped_at,
             'scraped_url': station['scraped_url'],
-            'loc': {
-                'type': 'Point',
-                'coordinates': [station['lon'], station['lat']]
-            }
+            'loc': {'type': 'Point', 'coordinates': [station['lon'], station['lat']]}
         },
-        "$set": {"last_update_date": datetime.utcnow()}
+        "$set": {
+            'prices': prices_dict,
+            "prices_last_hour": {station: {str(minute): price} for station, price in prices_dict.items()},
+            "prices_last_24h": {station: {str(hour): price} for station, price in prices_dict.items()},
+            "prices_last_yday": {station: {str(day_of_the_year): price} for station, price in prices_dict.items()},
+            "updated_at": datetime.utcnow(),
+        }
     }, upsert=True)
 
-    db['stations']
-
-
-"""
-def save_station_x(station):
-    key = station['key']
-    name = station['name']
-    scraped_at = station['scraped_at']
-    timestamp = int(datetime.strptime(scraped_at, '%Y-%m-%d %H:%M:%S').timestamp())
-
-    # pprint(station)
-
-    # 1. Current / Meta
-    r.hmset('stations:meta:%s' % key, flatten_dict({
-        'key': key,
-        'name': name,
-        'address': station['address'],
-        'xcode': station['xcode'],
-        'xid': station['xid'],
-        'prices': {FUEL_CODES[k]: v for k, v in station['prices'].items()},
-        'lat': station['lat'],
-        'lon': station['lon'],
-        'scraped_at': timestamp,
-        'scraped_url': station['scraped_url']
-    }, separator=":"))
-
-    # Prices
-    for price_name, value in station['prices'].items():
-        price_code = FUEL_CODES[price_name]
-        r.zadd("stations:prices:%s:%s" % (key, price_code), value, timestamp)
-
-    # Price at location
-    r.execute_command("GEOADD", "locations", station['lat'], station['lon'], key)
-"""
+    return station
 
 
 def fix_image_path(path):
